+++ packages/ink-contracts/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-### Changed
+## Unreleased
+
+### Added
+
+- ink-client: Add support for nested storage roots.
 
 ## 0.1.0 - 2024-10-12