+++ packages/ink-contracts/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - ink-client: Add support for nested storage roots.
 
+### Changed
+
+- Dynamic builder and client accept event topics for event decoding.
+
 ## 0.1.0 - 2024-10-12
 
 Initial release